chore: client & swap factory folder structure
