+++ src/factory/swap/index.ts
@@ -1,15 +1,15 @@
-import buildConfig from '../build.config';
-import { Network, SwapProviderType } from '../store/types';
-import { AstroportSwapProvider } from '../swaps/astroport/AstroportSwapProvider';
-import { FastbtcSwapProvider } from '../swaps/fastbtc/FastbtcSwapProvider';
-import { LiqualitySwapProvider } from '../swaps/liquality/LiqualitySwapProvider';
-import { LiqualityBoostERC20toNative } from '../swaps/liqualityboost/liqualityBoostERC20toNative/LiqualityBoostERC20toNative';
-import { LiqualityBoostNativeToERC20 } from '../swaps/liqualityboost/liqualityBoostNativeToERC20/LiqualityBoostNativeToERC20';
-import { OneinchSwapProvider } from '../swaps/oneinch/OneinchSwapProvider';
-import { SovrynSwapProvider } from '../swaps/sovryn/SovrynSwapProvider';
-import { SwapProvider } from '../swaps/SwapProvider';
-import { ThorchainSwapProvider } from '../swaps/thorchain/ThorchainSwapProvider';
-import { UniswapSwapProvider } from '../swaps/uniswap/UniswapSwapProvider';
+import buildConfig from '../../build.config';
+import { Network, SwapProviderType } from '../../store/types';
+import { AstroportSwapProvider } from '../../swaps/astroport/AstroportSwapProvider';
+import { FastbtcSwapProvider } from '../../swaps/fastbtc/FastbtcSwapProvider';
+import { LiqualitySwapProvider } from '../../swaps/liquality/LiqualitySwapProvider';
+import { LiqualityBoostERC20toNative } from '../../swaps/liqualityboost/liqualityBoostERC20toNative/LiqualityBoostERC20toNative';
+import { LiqualityBoostNativeToERC20 } from '../../swaps/liqualityboost/liqualityBoostNativeToERC20/LiqualityBoostNativeToERC20';
+import { OneinchSwapProvider } from '../../swaps/oneinch/OneinchSwapProvider';
+import { SovrynSwapProvider } from '../../swaps/sovryn/SovrynSwapProvider';
+import { SwapProvider } from '../../swaps/SwapProvider';
+import { ThorchainSwapProvider } from '../../swaps/thorchain/ThorchainSwapProvider';
+import { UniswapSwapProvider } from '../../swaps/uniswap/UniswapSwapProvider';
 
 const providers = {
   [SwapProviderType.Liquality]: LiqualitySwapProvider,
