+++ src/swaps/liqualityboost/liqualityBoostERC20toNative/LiqualityBoostERC20toNative.ts
@@ -1,6 +1,6 @@
 import { assets, unitToCurrency } from '@liquality/cryptoassets';
 import BN from 'bignumber.js';
-import { getSwapProvider } from '../../../factory/swapProvider';
+import { getSwapProvider } from '../../../factory/swap';
 import { ActionContext } from '../../../store';
 import { withInterval } from '../../../store/actions/performNextAction/utils';
 import { Asset, Network, SwapHistoryItem, WalletId } from '../../../store/types';
